backend/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.7.4
pydantic-settings==2.5.2

# LangChain
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-chroma==0.2.3
langchain-pinecone==0.2.0
langchain-huggingface==0.1.2
langchain-google-genai==2.1.9
langchain-google-community==2.0.7

# Vector Stores
chromadb==0.5.3
pinecone-client==5.0.1

# Embeddings
sentence-transformers==3.3.0
huggingface-hub==0.23.0
transformers==4.41.2
torch==2.3.0

# Document Processing
pypdf==3.17.4
beautifulsoup4==4.12.2
lxml==4.9.3

# Google
google-search-results==2.4.2
google-api-python-client==2.179.0

# Cloud Storage
boto3==1.34.0
google-cloud-storage==2.17.0

# Monitoring
prometheus-fastapi-instrumentator==6.1.0

# Server
gunicorn==23.0.0

# Utilities
requests==2.32.0
python-dotenv==1.0.1